fix: split HistoryManagement into multiple files
