--- composeApp/src/androidMain/kotlin/com/eren76/mangly/composables/screens/HistoryManagement.kt
@@ -1,389 +1,0 @@
-package com.eren76.mangly.composables.screens
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHostController
-import coil3.compose.SubcomposeAsyncImage
-import com.eren76.mangly.composables.shared.image.CoverCache
-import com.eren76.mangly.composables.shared.image.CoverImageRequests
-import com.eren76.mangly.composables.shared.image.ImageLoadingComposable
-import com.eren76.mangly.composables.shared.image.ImageLoadingErrorComposable
-import com.eren76.mangly.rooms.entities.HistoryChapterEntity
-import com.eren76.mangly.rooms.entities.HistoryEntity
-import com.eren76.mangly.rooms.entities.HistoryWithReadChapters
-import com.eren76.mangly.viewmodels.ExtensionMetadataViewModel
-import com.eren76.mangly.viewmodels.HistoryViewModel
-import com.eren76.manglyextension.plugins.ExtensionMetadata
-import com.eren76.manglyextension.plugins.Source
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import java.io.File
-import java.net.URLEncoder
-import java.text.DateFormat
-import java.util.Date
-import java.util.SortedMap
-import java.util.TreeMap
-import java.util.UUID
-
-@Composable
-fun HistoryManagement(
-    historyViewModel: HistoryViewModel,
-    navHostController: NavHostController,
-    extensionMetaDataViewModel: ExtensionMetadataViewModel
-) {
-    val historyWithChapters by historyViewModel.historyWithChapters
-
-    if (historyWithChapters.all { it.readChapters.isEmpty() }) {
-        HistoryEmptyState()
-        return
-    }
-
-    val historyData: SortedMap<Long, HistoryEntity> = remember(historyWithChapters) {
-        getHistoryDataNewestToOldest(historyViewModel)
-    }
-
-    val sourcesById: Map<UUID, ExtensionMetadata> =
-        remember(extensionMetaDataViewModel.getAllSources()) {
-            extensionMetaDataViewModel.getAllSources().associateBy { metadata ->
-                UUID.fromString(metadata.source.getExtensionId())
-            }
-        }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 16.dp, vertical = 8.dp)
-    ) {
-        Text(
-            text = "History",
-            style = MaterialTheme.typography.headlineSmall,
-            modifier = Modifier.padding(vertical = 8.dp)
-        )
-
-        HorizontalDivider()
-
-        LazyColumn(
-            modifier = Modifier.fillMaxSize(),
-            contentPadding = PaddingValues(vertical = 12.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            items(
-                items = historyData.entries.toList(),
-                key = { entry ->
-                    val timeKey = entry.key
-                    val historyEntity = entry.value
-                    historyEntity.id.toString() + timeKey.toString()
-                }
-            ) { entry ->
-                val timeKey = entry.key
-                val historyEntity = entry.value
-
-                val sourceMetadata = sourcesById[historyEntity.extensionId]
-                val source: Source? = sourceMetadata?.source
-
-                HistoryRow(
-                    historyEntity = historyEntity,
-                    lastReadAt = timeKey,
-                    source = source,
-                    historyViewModel = historyViewModel,
-                    onClick = {
-                        val encodedUrl = URLEncoder.encode(
-                            historyEntity.mangaUrl,
-                            Charsets.UTF_8.name()
-                        )
-                        extensionMetaDataViewModel.setSelectedSource(sourcesById[historyEntity.extensionId]!!) // TODO: not ideal
-                        navHostController.navigate("chapters/$encodedUrl")
-                    }
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun HistoryEmptyState() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-            Text(
-                text = "No history yet",
-                style = MaterialTheme.typography.titleMedium
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-            Text(
-                text = "Start reading a manga to see it here.",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-}
-
-@Composable
-private fun HistoryRow(
-    historyEntity: HistoryEntity,
-    lastReadAt: Long,
-    source: Source?,
-    historyViewModel: HistoryViewModel,
-    onClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 4.dp)
-            .clickable(onClick = onClick),
-        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            HistoryCoverImage(
-                historyEntity = historyEntity,
-                source = source,
-                historyViewModel = historyViewModel,
-                modifier = Modifier
-                    .size(110.dp)
-            )
-
-            Spacer(modifier = Modifier.width(16.dp))
-
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = historyEntity.mangaName,
-                    style = MaterialTheme.typography.titleMedium,
-                    maxLines = 3,
-                    overflow = TextOverflow.Ellipsis
-                )
-
-                Spacer(modifier = Modifier.height(6.dp))
-
-                Text(
-                    text = "Last read: " + formatLastRead(lastReadAt),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun HistoryCoverImage(
-    historyEntity: HistoryEntity,
-    source: Source?,
-    historyViewModel: HistoryViewModel,
-    modifier: Modifier = Modifier
-) {
-    if (source == null) {
-        Box(
-            modifier = modifier,
-            contentAlignment = Alignment.Center
-        ) {
-            CircularProgressIndicator(modifier = Modifier.size(16.dp))
-        }
-        return
-    }
-
-    HistoryCoverImageLoader(
-        historyEntity = historyEntity,
-        source = source,
-        historyViewModel = historyViewModel,
-        modifier = modifier
-    )
-}
-
-@Composable
-private fun HistoryCoverImageLoader(
-    historyEntity: HistoryEntity,
-    source: Source,
-    historyViewModel: HistoryViewModel,
-    modifier: Modifier = Modifier
-) {
-    val context = LocalContext.current
-
-    var localCoverFile by remember(historyEntity.id, historyEntity.coverImageFilename) {
-        mutableStateOf<File?>(null)
-    }
-
-    LaunchedEffect(historyEntity.id, historyEntity.coverImageFilename) {
-        localCoverFile = historyEntity.coverImageFilename?.let { filename ->
-            historyViewModel.getCoverFile(filename = filename, context = context)
-        }
-    }
-
-    var imageForList by remember(historyEntity.id, historyEntity.mangaUrl) {
-        mutableStateOf<Source.ImageForChaptersList?>(null)
-    }
-    var loadFailed by remember(historyEntity.id, historyEntity.mangaUrl) { mutableStateOf(false) }
-
-    LaunchedEffect(historyEntity.id, historyEntity.mangaUrl, localCoverFile) {
-        if (localCoverFile != null) return@LaunchedEffect
-
-        loadFailed = false
-        imageForList = null
-
-        val coverInfo: Source.ImageForChaptersList? = runCatching {
-            withContext(Dispatchers.IO) { source.getImageForChaptersList(historyEntity.mangaUrl) }
-        }.getOrNull()
-
-        if (coverInfo == null || coverInfo.imageUrl.isBlank()) {
-            loadFailed = true
-            return@LaunchedEffect
-        }
-
-        imageForList = coverInfo
-
-        val downloaded = runCatching {
-            CoverCache.downloadImage(coverInfo.imageUrl, coverInfo.headers)
-        }.getOrNull()
-
-        if (downloaded == null || downloaded.bytes.isEmpty()) {
-            loadFailed = true
-            return@LaunchedEffect
-        }
-
-        val ext = CoverCache.inferImageExtension(
-            contentType = downloaded.contentType,
-            finalUrl = downloaded.finalUrl,
-            originalUrl = coverInfo.imageUrl
-        )
-
-        val filename = "${historyEntity.id}.$ext"
-
-        val savedFile = withContext(Dispatchers.IO) {
-            historyViewModel.saveCoverBytes(
-                filename = filename,
-                bytes = downloaded.bytes,
-                context = context
-            )
-        }
-
-        historyViewModel.updateHistoryCoverFilename(
-            historyId = historyEntity.id,
-            filename = filename
-        )
-
-        localCoverFile = savedFile
-    }
-
-    val localRequest = remember(localCoverFile) {
-        CoverImageRequests.local(context = context, file = localCoverFile)
-    }
-
-    val networkHeaders = remember(imageForList?.headers) {
-        CoverCache.buildNetworkHeaders(imageForList?.headers ?: emptyList())
-    }
-
-    val remoteRequest = remember(imageForList?.imageUrl, networkHeaders) {
-        CoverImageRequests.remote(
-            context = context,
-            imageForList = imageForList,
-            networkHeaders = networkHeaders,
-            crossfade = true
-        )
-    }
-
-    when {
-        loadFailed -> {
-            Box(modifier = modifier, contentAlignment = Alignment.Center) {
-                ImageLoadingErrorComposable()
-            }
-        }
-
-        localRequest != null -> {
-            SubcomposeAsyncImage(
-                model = localRequest,
-                contentDescription = historyEntityDescriptionFromUrl(historyEntity.mangaUrl),
-                modifier = modifier,
-                contentScale = ContentScale.Crop,
-                loading = { ImageLoadingComposable() },
-                error = { ImageLoadingErrorComposable() }
-            )
-        }
-
-        imageForList?.imageUrl == null -> {
-            Box(modifier = modifier, contentAlignment = Alignment.Center) {
-                ImageLoadingComposable()
-            }
-        }
-
-        else -> {
-            SubcomposeAsyncImage(
-                model = remoteRequest,
-                contentDescription = historyEntityDescriptionFromUrl(historyEntity.mangaUrl),
-                modifier = modifier,
-                contentScale = ContentScale.Crop,
-                loading = { ImageLoadingComposable() },
-                error = { ImageLoadingErrorComposable() }
-            )
-        }
-    }
-}
-
-private fun historyEntityDescriptionFromUrl(url: String): String = "Cover for $url"
-
-fun getHistoryDataNewestToOldest(historyViewModel: HistoryViewModel): SortedMap<Long, HistoryEntity> {
-    val historyEntitiesNewestToOldestReadState: SortedMap<Long, HistoryEntity> =
-        TreeMap(compareByDescending { it })
-    val historyEntityByTime = HashMap<Long, HistoryEntity>()
-
-    for (chapter: HistoryWithReadChapters in historyViewModel.historyWithChapters.value) {
-        var latest = 0L
-        for (readChapter: HistoryChapterEntity in chapter.readChapters) {
-            val readAt = readChapter.readAt ?: continue
-            if (readAt > latest) {
-                latest = readAt
-            }
-        }
-        if (latest > 0L) {
-            historyEntityByTime[latest] = chapter.history
-        }
-    }
-
-    for ((key, historyEntity) in historyEntityByTime) {
-        historyEntitiesNewestToOldestReadState[key] = historyEntity
-    }
-
-    return historyEntitiesNewestToOldestReadState
-}
-
-fun formatLastRead(timestamp: Long): String {
-    return DateFormat.getDateTimeInstance(DateFormat.MEDIUM, DateFormat.SHORT)
-        .format(Date(timestamp))
-}
